Move test_recursion into test_basic.
